分类器模型启用GPU加速，Haar特征启用向量化计算，大幅提高速度
分类器模型启用GPU加速，Haar特征启用向量化计算，且发现特征在CPU上比在GPU上计算速度更快
